+++ paymentsheet/src/test/java/com/stripe/android/paymentsheet/ui/UpdatePaymentMethodUITest.kt
@@ -4,11 +4,9 @@ import android.os.Build
 import android.os.Parcel
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.test.assertAll
-import androidx.compose.ui.test.assertContentDescriptionEquals
 import androidx.compose.ui.test.assertIsDisplayed
 import androidx.compose.ui.test.assertIsEnabled
 import androidx.compose.ui.test.assertIsNotEnabled
-import androidx.compose.ui.test.assertTextContains
 import androidx.compose.ui.test.assertTextEquals
 import androidx.compose.ui.test.isEnabled
 import androidx.compose.ui.test.isNotEnabled
@@ -42,113 +40,6 @@ class UpdatePaymentMethodUITest {
 
     @get:Rule
     val composeRule = createComposeRule()
-
-    @Test
-    fun missingExpiryDate_displaysDots() {
-        val card = PaymentMethodFixtures.CARD_PAYMENT_METHOD.copy(
-            card = PaymentMethodFixtures.CARD_PAYMENT_METHOD.card!!.copy(
-                expiryMonth = null,
-            )
-        )
-
-        runScenario(displayableSavedPaymentMethod = card.toDisplayableSavedPaymentMethod()) {
-            assertExpiryDateEquals(
-                "••/••"
-            )
-        }
-    }
-
-    @Test
-    fun invalidExpiryMonth_displaysDots() {
-        val card = PaymentMethodFixtures.CARD_PAYMENT_METHOD.copy(
-            card = PaymentMethodFixtures.CARD_PAYMENT_METHOD.card!!.copy(
-                expiryMonth = -1,
-            )
-        )
-
-        runScenario(displayableSavedPaymentMethod = card.toDisplayableSavedPaymentMethod()) {
-            assertExpiryDateEquals(
-                "••/••"
-            )
-        }
-    }
-
-    @Test
-    fun invalidExpiryYear_displaysDots() {
-        val card = PaymentMethodFixtures.CARD_PAYMENT_METHOD.copy(
-            card = PaymentMethodFixtures.CARD_PAYMENT_METHOD.card!!.copy(
-                expiryYear = 202,
-            )
-        )
-
-        runScenario(displayableSavedPaymentMethod = card.toDisplayableSavedPaymentMethod()) {
-            assertExpiryDateEquals(
-                "••/••"
-            )
-        }
-    }
-
-    @Test
-    fun singleDigitExpiryMonth_hasLeadingZero() {
-        val card = PaymentMethodFixtures.CARD_PAYMENT_METHOD.copy(
-            card = PaymentMethodFixtures.CARD_PAYMENT_METHOD.card!!.copy(
-                expiryMonth = 8,
-                expiryYear = 2029,
-            )
-        )
-
-        runScenario(displayableSavedPaymentMethod = card.toDisplayableSavedPaymentMethod()) {
-            assertExpiryDateEquals(
-                "08/29"
-            )
-        }
-    }
-
-    @Test
-    fun doubleDigitExpiryMonth_doesNotHaveLeadingZero() {
-        val card = PaymentMethodFixtures.CARD_PAYMENT_METHOD.copy(
-            card = PaymentMethodFixtures.CARD_PAYMENT_METHOD.card!!.copy(
-                expiryMonth = 11,
-                expiryYear = 2029,
-            )
-        )
-
-        runScenario(displayableSavedPaymentMethod = card.toDisplayableSavedPaymentMethod()) {
-            assertExpiryDateEquals(
-                "11/29"
-            )
-        }
-    }
-
-    @Test
-    fun threeDigitCvcCardBrand_displaysThreeDotsForCvc() {
-        val card = PaymentMethodFixtures.CARD_PAYMENT_METHOD.copy(
-            card = PaymentMethodFixtures.CARD_PAYMENT_METHOD.card!!.copy(
-                brand = CardBrand.Visa
-            )
-        )
-
-        runScenario(displayableSavedPaymentMethod = card.toDisplayableSavedPaymentMethod()) {
-            assertCvcEquals(
-                "•••"
-            )
-        }
-    }
-
-    @Test
-    fun fourDigitCvcCardBrand_displaysFourDotsForCvc() {
-        val card = PaymentMethodFixtures.CARD_PAYMENT_METHOD.copy(
-            card = PaymentMethodFixtures.CARD_PAYMENT_METHOD.card!!.copy(
-                brand = CardBrand.AmericanExpress
-            )
-        )
-
-        runScenario(displayableSavedPaymentMethod = card.toDisplayableSavedPaymentMethod()) {
-            assertCvcEquals(
-                "••••"
-            )
-        }
-    }
 
     @Test
     fun canRemoveIsFalse_removeButtonHidden() = runScenario(
@@ -349,43 +240,10 @@ class UpdatePaymentMethodUITest {
     }
 
     @Test
-    fun modifiableCard_cbcDropdownIsShown() {
-        runScenario(
-            displayableSavedPaymentMethod = PaymentMethodFixtures
-                .CARD_WITH_NETWORKS_PAYMENT_METHOD
-                .toDisplayableSavedPaymentMethod()
-        ) {
-            composeRule.onNodeWithTag(DROPDOWN_MENU_CLICKABLE_TEST_TAG).assertExists()
-        }
-    }
-
-    @Test
-    fun notModifiableCard_cbcDropdownIsNotShown() {
-        runScenario(
-            displayableSavedPaymentMethod = PaymentMethodFixtures.displayableCard()
-        ) {
-            composeRule.onNodeWithTag(DROPDOWN_MENU_CLICKABLE_TEST_TAG).assertDoesNotExist()
-        }
-    }
-
-    @Test
     fun bankAccount_cbcDropdownIsNotShown() {
         runScenario(
             displayableSavedPaymentMethod = PaymentMethodFixtures.US_BANK_ACCOUNT.toDisplayableSavedPaymentMethod()
         ) {
-            composeRule.onNodeWithTag(DROPDOWN_MENU_CLICKABLE_TEST_TAG).assertDoesNotExist()
-        }
-    }
-
-    @Test
-    fun isNotModifiablePM_cbcDropdownIsNotShown() {
-        runScenario(
-            displayableSavedPaymentMethod = PaymentMethodFixtures
-                .CARD_WITH_NETWORKS_PAYMENT_METHOD
-                .toDisplayableSavedPaymentMethod(),
-            isModifiablePaymentMethod = false,
-        ) {
-            // Even if the card itself is modifiable, if we set isModifiablePaymentMethod to false, CBC is hidden.
             composeRule.onNodeWithTag(DROPDOWN_MENU_CLICKABLE_TEST_TAG).assertDoesNotExist()
         }
     }
@@ -408,19 +266,6 @@ class UpdatePaymentMethodUITest {
                 )
             )
             assertThat(viewActionRecorder.viewActions).isEmpty()
-        }
-    }
-
-    @Test
-    fun `Card drop down has accessibility label`() {
-        runScenario(
-            displayableSavedPaymentMethod = PaymentMethodFixtures
-                .CARD_WITH_NETWORKS_PAYMENT_METHOD
-                .toDisplayableSavedPaymentMethod(),
-            initialCardBrand = CardBrand.Visa,
-        ) {
-            composeRule.onNodeWithTag(DROPDOWN_MENU_CLICKABLE_TEST_TAG)
-                .assertContentDescriptionEquals("Visa")
         }
     }
 
@@ -477,18 +322,6 @@ class UpdatePaymentMethodUITest {
             )
             assertThat(viewActionRecorder.viewActions).isEmpty()
         }
-    }
-
-    private fun assertExpiryDateEquals(text: String) {
-        composeRule.onNodeWithTag(UPDATE_PM_EXPIRY_FIELD_TEST_TAG).assertTextContains(
-            text
-        )
-    }
-
-    private fun assertCvcEquals(text: String) {
-        composeRule.onNodeWithTag(UPDATE_PM_CVC_FIELD_TEST_TAG).assertTextContains(
-            text
-        )
     }
 
     private fun runScenario(
